+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Removed
 
 ### Fixed
+- Frame concurrency 문제 해결
 
 ### Security